python/simple_ecu/requirements.txt
requests==2.21.0
remotivelabs-broker
